testing/web-platform/tests/dom/nodes/Document-getElementsByTagNameNS.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Document.getElementsByTagNameNS</title>
<link rel=help href="https://dom.spec.whatwg.org/#dom-document-getelementsbytagnamens">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
// TODO: Actual tests; wildcards.
test(function() {
  assert_false(document.getElementsByTagNameNS("http://www.w3.org/1999/xhtml", "html") instanceof NodeList, "NodeList")
  assert_true(document.getElementsByTagNameNS("http://www.w3.org/1999/xhtml", "html") instanceof HTMLCollection, "HTMLCollection")
  var firstCollection = document.getElementsByTagNameNS("http://www.w3.org/1999/xhtml", "html"),
      secondCollection = document.getElementsByTagNameNS("http://www.w3.org/1999/xhtml", "html")
  assert_true(firstCollection !== secondCollection || firstCollection === secondCollection,
              "Caching is allowed.")
})

test(function() {
  var t = document.body.appendChild(document.createElementNS("test", "body"))
  this.add_cleanup(function() {document.body.removeChild(t)})
  var actual = document.getElementsByTagNameNS("*", "body");
  var expected = [];
  var get_elements = function(node) {
    for (var i = 0; i < node.childNodes.length; i++) {
      var child = node.childNodes[i];
      if (child.nodeType === child.ELEMENT_NODE) {
        if (child.localName == "body") {
          expected.push(child);
        }
        get_elements(child);
      }
    }
  }
  get_elements(document);
  assert_array_equals(actual, expected);
}, "getElementsByTagNameNS('*', 'body')")

test(function() {
  assert_array_equals(document.getElementsByTagNameNS("", "*"), []);
  var t = document.body.appendChild(document.createElementNS("", "body"))
  this.add_cleanup(function() {document.body.removeChild(t)})
  assert_array_equals(document.getElementsByTagNameNS("", "*"), [t]);
}, "Empty string namespace")

test(function() {
  var t = document.body.appendChild(document.createElementNS("test", "body"))
  this.add_cleanup(function() {document.body.removeChild(t)})
  assert_array_equals(document.getElementsByTagNameNS("test", "body"), [t]);
}, "body element in test namespace, no prefix")

test(function() {
  var t = document.body.appendChild(document.createElementNS("test", "test:body"))
  this.add_cleanup(function() {document.body.removeChild(t)})
  assert_array_equals(document.getElementsByTagNameNS("test", "body"), [t]);
}, "body element in test namespace, prefix")

test(function() {
  var t = document.body.appendChild(document.createElementNS("test", "BODY"))
  this.add_cleanup(function() {document.body.removeChild(t)})
  assert_array_equals(document.getElementsByTagNameNS("test", "BODY"), [t]);
  assert_array_equals(document.getElementsByTagNameNS("test", "body"), []);
}, "BODY element in test namespace, no prefix")

test(function() {
  var t = document.body.appendChild(document.createElementNS("http://www.w3.org/1999/xhtml", "abc"))
  this.add_cleanup(function() {document.body.removeChild(t)})
  assert_array_equals(document.getElementsByTagNameNS("http://www.w3.org/1999/xhtml", "abc"), [t]);
 assert_array_equals(document.getElementsByTagNameNS("http://www.w3.org/1999/xhtml", "ABC"), []);
  assert_array_equals(document.getElementsByTagNameNS("test", "ABC"), []);
}, "abc element in html namespace")

test(function() {
  var t = document.body.appendChild(document.createElementNS("http://www.w3.org/1999/xhtml", "ABC"))
  this.add_cleanup(function() {document.body.removeChild(t)})
  assert_array_equals(document.getElementsByTagNameNS("http://www.w3.org/1999/xhtml", "abc"), []);
  assert_array_equals(document.getElementsByTagNameNS("http://www.w3.org/1999/xhtml", "ABC"), [t]);
}, "ABC element in html namespace")

test(function() {
  var t = document.body.appendChild(document.createElementNS("http://www.w3.org/1999/xhtml", "AÇ"))
  this.add_cleanup(function() {document.body.removeChild(t)})
  assert_array_equals(document.getElementsByTagNameNS("http://www.w3.org/1999/xhtml", "AÇ"), [t]);
  assert_array_equals(document.getElementsByTagNameNS("test", "aÇ"), []);
  assert_array_equals(document.getElementsByTagNameNS("test", "aç"), []);
}, "AÇ, case sensitivity")

test(function() {
  var t = document.body.appendChild(document.createElementNS("test", "test:BODY"))
  this.add_cleanup(function() {document.body.removeChild(t)})
  assert_array_equals(document.getElementsByTagNameNS("test", "BODY"), [t]);
  assert_array_equals(document.getElementsByTagNameNS("test", "body"), []);
}, "BODY element in test namespace, prefix")

test(function() {
  var t = document.body.appendChild(document.createElementNS("test", "test:test"))
  this.add_cleanup(function() {document.body.removeChild(t)})
  var actual = document.getElementsByTagNameNS("http://www.w3.org/1999/xhtml", "*");
  var expected = [];
  var get_elements = function(node) {
    for (var i = 0; i < node.childNodes.length; i++) {
      var child = node.childNodes[i];
      if (child.nodeType === child.ELEMENT_NODE) {
        if (child !== t) {
          expected.push(child);
        }
        get_elements(child);
      }
    }
  }
  get_elements(document);
  assert_array_equals(actual, expected);
}, "getElementsByTagNameNS('http://www.w3.org/1999/xhtml', '*')")

test(function() {
  var actual = document.getElementsByTagNameNS("*", "*");
  var expected = [];
  var get_elements = function(node) {
    for (var i = 0; i < node.childNodes.length; i++) {
      var child = node.childNodes[i];
      if (child.nodeType === child.ELEMENT_NODE) {
        expected.push(child);
        get_elements(child);
      }
    }
  }
  get_elements(document);
  assert_array_equals(actual, expected);
}, "getElementsByTagNameNS('*', '*')")

test(function() {
  assert_array_equals(document.getElementsByTagNameNS("**", "*"), []);
  assert_array_equals(document.getElementsByTagNameNS(null, "0"), []);
  assert_array_equals(document.getElementsByTagNameNS(null, "div"), []);
}, "Empty lists")
</script>
